renamed FormContainer to 'MappingForm'
